[3439] [Time Beats: 100.00%] [Memory Beats: 59.76%] - LeetPush

+++ LeetCode/Reschedule-Meetings-for-Maximum-Free-Time-I.cpp
@@ -3,13 +3,13 @@ public:
     int maxFreeTime(int eventTime, int k, vector<int>& startTime, vector<int>& endTime) {
         int n = startTime.size();
         vector<int> gaps(n + 1);
-        gaps[0] = (int)startTime[0] - 0;
+        gaps[0] = startTime[0] - 0;
 
         for (int i = 1; i < n; ++i) {
-            gaps[i] = (int)startTime[i] - endTime[i - 1];
+            gaps[i] = startTime[i] - endTime[i - 1];
         }
 
-        gaps[n] = (int)eventTime - endTime[n - 1];
+        gaps[n] = eventTime - endTime[n - 1];
 
         vector<int> prefix(n + 2, 0);
         for (int i = 1; i <= n + 1; ++i) {
